fullstack-app/python-backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
redis==5.0.8
gunicorn==22.0.0
